+++ Code/JavaWrappers/gmwrapper/src-test/org/RDKit/SubstanceGroupTests.java
@@ -50,7 +50,7 @@ public class SubstanceGroupTests extends GraphMolTest {
         assertEquals(0, supGroup.getStringVectProp("DATAFIELDS").size());
         assertEquals(1, supGroup.getAtoms().size());
         assertEquals(0, supGroup.getBonds().size());
-        assertEquals((Integer)1, supGroup.getAtoms().get(0));
+        assertEquals((long)1,(long) supGroup.getAtoms().get(0));
     }
 
     @Test
@@ -66,11 +66,11 @@ public class SubstanceGroupTests extends GraphMolTest {
         assertTrue(substanceGroup.hasProp("index"));
         assertEquals(2, substanceGroup.getUIntProp("index") );
         assertEquals(2, substanceGroup.getAtoms().size());
-        assertEquals((Integer)3, substanceGroup.getAtoms().get(0));
-        assertEquals((Integer)2, substanceGroup.getAtoms().get(1));
+        assertEquals((long)3,(long) substanceGroup.getAtoms().get(0));
+        assertEquals((long)2,(long) substanceGroup.getAtoms().get(1));
         assertEquals(2, substanceGroup.getBonds().size());
-        assertEquals((Integer)1, substanceGroup.getBonds().get(0));
-        assertEquals((Integer)3, substanceGroup.getBonds().get(1));
+        assertEquals((long)1,(long) substanceGroup.getBonds().get(0));
+        assertEquals((long)3,(long) substanceGroup.getBonds().get(1));
         assertTrue(substanceGroup.hasProp("PARENT"));
         assertEquals(10, substanceGroup.getUIntProp("PARENT"));
 
@@ -78,14 +78,14 @@ public class SubstanceGroupTests extends GraphMolTest {
         assertTrue(substanceGroup.hasProp("index"));
         assertEquals(10, substanceGroup.getUIntProp("index") );
         assertEquals(5, substanceGroup.getAtoms().size());
-        assertEquals((Integer)3, substanceGroup.getAtoms().get(0));
-        assertEquals((Integer)2, substanceGroup.getAtoms().get(1));
-        assertEquals((Integer)4, substanceGroup.getAtoms().get(2));
-        assertEquals((Integer)5, substanceGroup.getAtoms().get(3));
-        assertEquals((Integer)7, substanceGroup.getAtoms().get(4));
+        assertEquals((long)3,(long) substanceGroup.getAtoms().get(0));
+        assertEquals((long)2,(long) substanceGroup.getAtoms().get(1));
+        assertEquals((long)4,(long) substanceGroup.getAtoms().get(2));
+        assertEquals((long)5,(long) substanceGroup.getAtoms().get(3));
+        assertEquals((long)7,(long) substanceGroup.getAtoms().get(4));
         assertEquals(2, substanceGroup.getBonds().size());
-        assertEquals((Integer)1, substanceGroup.getBonds().get(0));
-        assertEquals((Integer) 5, substanceGroup.getBonds().get(1));
+        assertEquals((long)1,(long) substanceGroup.getBonds().get(0));
+        assertEquals((long)5,(long) substanceGroup.getBonds().get(1));
         assertFalse(substanceGroup.hasProp("PARENT"));
 
     }
